Task 6

+++ 0x00-hello_world/6-size.c
@@ -15,10 +15,10 @@ int main(void)
 	long long lli;
 	float f;
 	
-	printf("Size of a char: %d byte(s)\n", sizeof(c));
-	printf("Size of an int: %d byte(s)\n", sizeof(i));
-	printf("Size of a long int: %1d bytes(s)\n", sizeof(li));
-	printf("Size of a long longe int: %1d bytes(s)\n", sizeof(lli);
-	printf("Size of a float: %1d byte(s)\n", sizeof(f));
+	printf("Size of a char: %ld byte(s)\n",(unsigned long) sizeof(c));
+	printf("Size of an int: %ld byte(s)\n",(unsigned long) sizeof(i));
+	printf("Size of a long int: %ld bytes(s)\n",(unsigned long) sizeof(li));
+	printf("Size of a long longe int: %ld bytes(s)\n",(unsigned long) sizeof(lli);
+	printf("Size of a float: %ld byte(s)\n",(unsigned long) sizeof(f));
 	retutn (0);
 }
